schema: add 4.11.2 to 4.11.3 systemvmtemplate upgrade path
